requirements.txt
boto3==1.26.137
psycopg2-binary==2.9.6
requests==2.31.0
